mau/requirements.txt
autograd==1.3
cycler==0.10.0
decorator==4.4.2
future==0.18.2
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-mnist==0.7
pytz==2021.1
PyWavelets==1.1.1
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
tifffile==2021.6.14
torch==1.9.0
torchdiffeq==0.2.2
torchvision==0.10.0
typing-extensions==3.10.0.0
